[chore] #115 타입 변경

+++ data-remote/category/src/main/java/org/sopt/dataremote/category/response/ResponseCategoryEntireDto.kt
@@ -26,7 +26,7 @@ data class ResponseCategoryEntireDto(
 internal fun ResponseCategoryEntireDto.ResponseCategoryDto.toCoreModel() = Category(
   categoryId = categoryId,
   categoryTitle = categoryTitle,
-  toastNum = toastNum,
+  toastNum = toastNum.toLong(),
 )
 
 internal fun ResponseCategoryEntireDto.toCoreModel() = CategoryList(
